+++ bee-api/bee-api-types/src/responses.rs
@@ -129,11 +129,11 @@ pub struct BaseTokenResponse {
 #[derive(Clone, Debug, Eq, PartialEq, Serialize, Deserialize)]
 pub struct RentStructureResponse {
     #[serde(rename = "vByteCost")]
-    pub v_byte_cost: u64,
+    pub v_byte_cost: u32,
     #[serde(rename = "vByteFactorKey")]
-    pub v_byte_factor_key: u64,
+    pub v_byte_factor_key: u8,
     #[serde(rename = "vByteFactorData")]
-    pub v_byte_factor_data: u64,
+    pub v_byte_factor_data: u8,
 }
 
 /// Returned in [`InfoResponse`].