+++ bee-block/src/rand/parents.rs
@@ -1,9 +1,10 @@
 // Copyright 2020-2021 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-use bee_block::parent::Parents;
-
-use crate::rand::{block::rand_block_ids, number::rand_number_range};
+use crate::{
+    parent::Parents,
+    rand::{block::rand_block_ids, number::rand_number_range},
+};
 
 /// Generates random parents.
 pub fn rand_parents() -> Parents {