+++ bee-storage/bee-storage-test/src/output_id_to_created_output.rs
@@ -1,13 +1,12 @@
 // Copyright 2020-2021 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-use bee_block::output::OutputId;
-use bee_ledger::types::CreatedOutput;
+use bee_block::{output::OutputId, rand::output::rand_output_id};
+use bee_ledger_types::{rand::output::rand_created_output, CreatedOutput};
 use bee_storage::{
     access::{AsIterator, Batch, BatchBuilder, Delete, Exist, Fetch, Insert, MultiFetch, Truncate},
     backend,
 };
-use bee_test::rand::output::{rand_created_output, rand_output_id};
 
 pub trait StorageBackend:
     backend::StorageBackend