+++ bee-storage/bee-storage-test/src/snapshot_info.rs
@@ -1,12 +1,11 @@
 // Copyright 2020-2021 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-use bee_ledger::types::snapshot::SnapshotInfo;
+use bee_ledger_types::{rand::snapshot::rand_snapshot_info, snapshot::SnapshotInfo};
 use bee_storage::{
     access::{AsIterator, Batch, BatchBuilder, Delete, Exist, Fetch, Insert, Truncate},
     backend,
 };
-use bee_test::rand::snapshot::rand_snapshot_info;
 
 pub trait StorageBackend:
     backend::StorageBackend