+++ bee-test/src/lib.rs
@@ -5,7 +5,5 @@
 
 #![deny(missing_docs, warnings)]
 
-/// A module providing random generation of types of testing purposes.
-pub mod rand;
 /// A module providing ternary utilities for testing purposes.
 pub mod ternary;